Problem 2: Derive the sentiment of each tweet
Problem 2 is the tweet_sentiment.py file
The strategy here is that for each tweet:

1. Remove @ mentions
2. Remove # signs, perods (.), commas (,), explamation points (!), and question marks (?)
3. encode in utf-8
4. Iterate through term from the sentiment file (AFINN-111.txt). Check if contained in tweet
5. If it is, add to the score

The sentiment file is only 2.5k words, and the input file I am testing is only 20 MB,
So this algorithm isn't that costly. There are also multi-word phrases
in the sentiment file, which makes splitting a little hairier. One option would
be to take the multi-word sentiments and check those explicitly while using
some sort of splitting pattern, but I'll see if this is 'good enough' right now