Problem 3: Derive the sentiment of new terms
'Derive the sentiment of new terms'. This problem has some scaffold in
term_sentiment.py, and that is where the solution is done.

I didn't read over the linked paper (probably should have), but my solution
takes a simple approach.

1. Use the same scoring strategy from the tweet_sentiment.py problem to calculate score
2. Split the 'normalized' text
3. Remove every word that is already in the sentiment file
4. Calculate what percentage of the score this word represents
5. Average the totals by key

Once again, just using the firth approach that popped up. One problem I noticed
is that most (if not all) of the scores are very low.