Problem 5: Which State is happiest?
Copied code from the other .py files, again. This uses the score computation
and the tweet 'cleaner' from the 2nd assignment. The location determining
is really... brute-forcy and not so great.

The algorithm basically just searches for String matches with a state
abbreviation or the actual state name. It basically checks a handful of
exact options. A better option would be to statically create geoboxes and use the
latitude/longitude to compute which state the geohash lands in.

I add all of these to a list if there is a score and location. Then, the average
is computed and the state with the highest is printed out